refactor: 이동 로직 책임분리
