+++ src/main/java/model/policy/PathPolicy.java
@@ -1,0 +1,8 @@
+package model.policy;
+
+import model.board.Board;
+import model.position.Position;
+
+public interface PathPolicy {
+    boolean check(Position pos, Board board);
+}
